Move properties and tags methods to separate clients
