+++ georocket-cli/src/main/java/io/georocket/commands/RemovePropertyCommand.java
@@ -88,7 +88,7 @@ public class RemovePropertyCommand extends AbstractGeoRocketCommand {
   public void doRun(String[] remainingArgs, InputReader in, PrintWriter out,
       Handler<Integer> handler) throws OptionParserException, IOException {
     GeoRocketClient client = createClient();
-    client.getStore().removeProperties(query, layer, properties, ar -> {
+    client.getProperties().removeProperties(query, layer, properties, ar -> {
       if (ar.failed()) {
         client.close();
         Throwable t = ar.cause();
